outputImage is now its own module
